django_notifications_2.0/backend/requirements.txt
asgiref==3.10.0
channels==4.3.1
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
sqlparse==0.5.3
